requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.13.2
pydantic==2.8.2
python-multipart==0.0.12
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
